extend an existing test case
