--- test/cases/overloads2.out.py
@@ -1,22 +1,0 @@
-from typing import overload
-
-
-class Selector():
-    @overload
-    def union(self, other: Self) -> Self: ...
-    @overload
-    def union(self, other: IntoExpr) -> Expr: ...
-    def union(self, other: IntoExpr) -> Self | Expr:  # noqa: D102
-        match other:
-            case Selector():
-                return self.into_selector()
-            case _:
-                return super().__or__(other)
-
-    @overload
-    def __or__(self, other: Self) -> Self: ...
-    @overload
-    def __or__(self, other: IntoExpr) -> Expr: ...
-    @override
-    def __or__(self, other: IntoExpr) -> Self | Expr:
-        return self.union(other)
